Implemented Barnes-Hut quadtree and algorithm
- Two galaxies are created within the screen with points spread exponentially around their center
- QuadTree is rebuilt every frame
- Velocity and Position computations are super simple ( eg. v = v +a*dt)

Galaxies are not stable at all right now, the sim starts and all particles begin to move away from center. gotta fix.

+++ GalaxySimulator_Threaded_COMP426/COMP426_Assignment1/Particle.cpp
@@ -1,13 +1,10 @@
 #include "Particle.h"
 
-
-
-Particle::Particle(int idx, float mass) : idx_(idx), mass_(mass)
+Particle::Particle(int idx, double mass) : idx_(idx), mass_(mass)
 {
-	pos_ = glm::vec2(0.0f, 0.0f);
-	vel_ = glm::vec2(0.0f, 0.0f);
-	acc_ = glm::vec2(0.0f, 0.0f);
-
+	pos_ = glm::vec2(0.0, 0.0);
+	vel_ = glm::vec2(0.0, 0.0);
+	acc_ = glm::vec2(0.0, 0.0);
 }
